[Breaking] use dbus for unit management (requires python-dbus)
It seems that session termination in systemd became quite aggressive.
Effectively any process spawned by uwsm is being hit with TERM signals before
it can do anything. This includes systemctl invocation that is intended to stop
wayland-wm@*.service unit upon receiving termination signals.
The only way to fix this seems to be something that I wanted to avoid: dbus.
Python interface to this monstrosity is relatively nice though, so I've replaced
a lot of systemctl calls with it along the way.