feat: -F for hardcoding resulting path, explicit ID for aux subcommands, wip #36
Now ID is always the first argument for prepare-env and exec in aux
The rest is optional full command line.
If command line is empty or its first argument (executable) is empty,
executable will be determined from ID as usual.

Executable is specified only if given as path or -F argument is given
for start, otherwise it is left empty to be (re)discovered by units.